+++ devicemanagement/conanfile.py
@@ -37,7 +37,7 @@ class DeviceManagementConan(ConanFile):
 
     def requirements(self):
         self.requires("deviceApi/[>=0.7.0 <1.0.0]")
-        self.requires("deviceLayer/[>=1.0.0 <2.0.0]")
+        self.requires("deviceLayer/[>=1.1.0 <2.0.0]")
         self.requires("hostUtils/[>=0.1.0 <1.0.0]")
 
     def validate(self):